fix bug

+++ app/src/main/java/com/github/tvbox/osc/player/controller/BaseController.java
@@ -390,7 +390,7 @@ public abstract class BaseController extends BaseVideoController implements Gest
         mSeekPosition = position;
     }
 
-    protected long getStep(){
+    protected int getStep(){
         return 5000;
     }
